Moved object helpers to utils package
