+++ apps/web/src/components/stats/table-row-stats.tsx
@@ -1,7 +1,7 @@
-import { getKeys } from '@utils/common';
 import { getShortStatName, getTotalBaseStat } from '@utils/pokemon';
 import type { IBaseStats } from 'contract';
 import { Typography } from 'ui';
+import { getKeys } from 'utils';
 
 interface TableRowStatsProps {
   stats: IBaseStats;
