+++ packages/utils/src/common/index.ts
@@ -1,2 +1,3 @@
 export * from './array';
+export * from './object';
 export * from './string';
